Refactor error handling and utility functions
